menambahkan access jwt

+++ src/pages/api/[...users].ts
@@ -1,3 +1,4 @@
+import jwt from "jsonwebtoken"
 import { deleteData, retriveData, updateData } from "@/libs/firebase/firebase_service"
 import { NextApiRequest, NextApiResponse } from "next"
 
@@ -15,36 +16,61 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
             data
         })
     } else if (req.method === "PATCH") {
-        const { id, data } = req.body
-        await updateData("users", id, data, (result: boolean) => {
-            if (result) {
-                res.status(200).json({
-                    status: true,
-                    statusCode: 200,
-                    message: "success",
+        const { users }: any = req.query
+        const { data } = req.body
+        const token: any = req.headers.authorization?.split(" ")[1]
+
+        jwt.verify(token, process.env.NEXT_AUTH_SECRET || "", async (err: any, decoded: any) => {
+            if (decoded && decoded.role === "admin") {
+                await updateData("users", users[1], data, (result: boolean) => {
+                    if (result) {
+                        res.status(200).json({
+                            status: true,
+                            statusCode: 200,
+                            message: "success",
+                        })
+                    } else {
+                        res.status(400).json({
+                            status: false,
+                            statusCode: 200,
+                            message: "failed",
+                        })
+                    }
                 })
             } else {
-                res.status(400).json({
+                res.status(403).json({
                     status: false,
-                    statusCode: 200,
-                    message: "failed",
+                    statusCode: 403,
+                    message: "access denied",
                 })
             }
         })
+
     } else if (req.method === "DELETE") {
         const { users }: any = req.query
-        await deleteData("users", users[1], (result: boolean) => {
-            if (result) {
-                res.status(200).json({
-                    status: true,
-                    statusCode: 200,
-                    message: "success",
+        const token: any = req.headers.authorization?.split(" ")[1]
+        jwt.verify(token, process.env.NEXT_AUTH_SECRET || "", async (err: any, decoded: any) => {
+            if (decoded && decoded.role === "admin") {
+                await deleteData("users", users[1], (result: boolean) => {
+                    if (result) {
+                        res.status(200).json({
+                            status: true,
+                            statusCode: 200,
+                            message: "success",
+                        })
+                    } else {
+                        res.status(400).json({
+                            status: false,
+                            statusCode: 400,
+                            message: "failed",
+                        })
+                    }
                 })
             } else {
-                res.status(400).json({
+                res.status(403).json({
                     status: false,
-                    statusCode: 200,
-                    message: "failed",
+                    statusCode: 403,
+                    message: "access denied",
                 })
             }
         })
